requirements.txt
python-dotenv==1.0.0
imapclient==2.2.0
beautifulsoup4==4.12.2
pandas==2.0.3
